remove extensionsbody comments
